Phase 2 (4/4): what the first Windows compile found, and rustfmt's own answer again
`3a1421c`'s run got past everything the earlier failures hid — `hanglock-core`, `hanglock-render`,
`hanglock-platform`, the app crate and all their tests now compile and lint clean on both Windows ABIs —
and `hanglock-win` came back with ten diagnostics. This is the first time any of batch 2's Win32 side has
been type-checked, and the misses are the expected family: names that were used and never imported, a lint
that does not exist yet, a field that was read and written but not declared.

- `panel.rs`: `RowId` used in `AnswerFn`, in `answer` and in two tests, never imported through
  `hanglock_platform::panel` — added to the `use`.
- `panel.rs` tests: `rows()` yields rows by value, and `row_height`/`box_of` take `&Row` — the two test
  call sites now borrow. (The three production call sites already had `&Row` in scope, which is why only
  the `lib test` target showed these.)
- `sys.rs`: `#[allow(clippy::usize_as_ptr)]` was a lint invented in Clippy 1.78; the toolchain is pinned to
  1.77.2, so the allow itself was the error (`unknown_lints`), and then `unused_attributes` on top of it.
  Removed, with the reason written where the allow was, and a note that the lint arrives with the day the
  pin moves. No suppression was needed: the cast is a plain `usize as *mut c_void`, and the 1.77 lint set
  leaves it alone.
- `sys.rs`: `VK_MENU` is an `i32`, because that is what `GetKeyState` takes; `KEY_DOWN_MASK` stays `i16`,
  because it is masked against the return, which is a `SHORT`. One constant, two types, both correct: the
  alternative is a `.into()` at every call site and a reader who cannot tell which one is the OS's type.
- `displays.rs`: my `monitors()` was inserted *above* the doc comment and `#[must_use]` that belonged to
  `enumerate`, leaving two attributes on one item and none on the other. The stray pair is gone;
  `enumerate` has its own doc below, which is the one that describes it.
- `window.rs`: `Host.last_cursor` existed in `new_runtime`'s literal and in `WM_MOUSEMOVE`, and did not
  exist in the struct. Declared, with the sentence that says why velocity is measured from the screen
  cursor rather than from the message pair.

Then rustfmt again on the code I authored last time — nine hunks, applied from its own `--check` output
again rather than guessed: `fn item(...)` joined at 95 chars, three `self.item(...)` calls broken because
their *argument lists* pass `fn_call_width = 60` even though the lines fit in 100 (that is the rule this
file keeps teaching me: the width of a call is not the width of its line), `set_visible`'s one-line
if/else, and `step_hang`'s argument if/else broken for the same 50-column reason. Verified by re-running the
60-column audit over every call in the two files and finding no single-line call whose args exceed it.

`.gitignore` gains `*.orig`/`*.rej`: `patch --fuzz` leaves backups, and one of them — a 272-line copy of
`model.rs` — got committed last time. Not rewritten out of history; deleted here, which is where it belongs
now.

+++ crates/hanglock-win/src/sys.rs
@@ -38,11 +38,12 @@ pub type HMODULE = *mut c_void;
 pub type HANDLE = *mut c_void;
 pub type HFONT = *mut c_void;
 
-/// The `hMenu` argument for a *child* window, which Win32 overloads as the control id. Nothing
-/// dereferences the pointer — the OS reads the number back out of it — so this is the one place in the
-/// crate that turns an integer into a pointer, and the one place that says why the lint about it is
-/// wrong here.
-#[allow(clippy::usize_as_ptr)]
+/// The `hMenu` argument for a *child* window, which Win32 overloads as the control id: nothing
+/// dereferences the pointer, because the OS reads the number back out of it. This is the one place in
+/// the crate that turns an integer into a pointer, which is why it is a named function with a comment
+/// rather than a cast at each `CreateWindowExW` call site. (Clippy grows a lint for exactly this cast in
+/// a later release than the pinned 1.77.2; when the pin moves, an `#[allow]` here will need a reason, not
+/// just a suppression.)
 #[must_use]
 pub const fn id_menu(id: usize) -> HMENU {
     id as *mut core::ffi::c_void
@@ -128,7 +129,7 @@ pub const IDOK: usize = 1;
 pub const IDCANCEL: usize = 2;
 /// `BN_CLICKED`, the `HIWORD(wParam)` of a button's `WM_COMMAND`.
 pub const BN_CLICKED: usize = 0;
-pub const VK_MENU: i16 = 0x12;
+pub const VK_MENU: i32 = 0x12;
 /// `GetKeyState`'s sign bit: "down right now", as opposed to the toggle bit, which is what Alt+Shift
 /// menus flip and is not what we ask.
 pub const KEY_DOWN_MASK: i16 = -0x8000;
